Update Swift, Clang, and LLVM
This compiles

print("hello world")

ok, but almost everything else crashes because of a function signature mismatch
when calling a non-throwing closure passed into a function that expects a throwing closure.

Swift passes an extra parameter to the closure if it thinks it can throw; on all other platforms
passing an extra parameter is harmless; on WebAssembly this is not allowed.